Feature/16 Board 서비스 로직 작성

+++ src/main/java/com/example/demo/domain/board/dto/response/BoardGetRes.java
@@ -11,7 +11,7 @@ public record BoardGetRes(
         String title,
         String content,
         UserGetRes writer,
-        Long likeCount,
+        Integer likeCount,
         LocalDateTime createdAt,
         LocalDateTime updatedAt
 ) {
